+++ lib/boruta/adapters/ecto/oauth_mapper.ex
@@ -10,11 +10,14 @@ defimpl Boruta.Ecto.OauthMapper, for: Boruta.Ecto.Token do
 
   alias Boruta.Oauth
   alias Boruta.Ecto
+  alias Boruta.Ecto.Clients
   alias Boruta.Ecto.OauthMapper
 
   def to_oauth_schema(%Ecto.Token{} = token) do
-    token = repo().preload(token, [:client])
-    client = OauthMapper.to_oauth_schema(token.client)
+    client = case Clients.get_client(token.client_id) do
+      %Oauth.Client{} = client -> client
+      _ -> nil
+    end
     resource_owner = token.resource_owner || with "" <> sub <- token.sub, # token is linked to a resource_owner
       {:ok, resource_owner} <- resource_owners().get_by(sub: sub) do
       resource_owner